requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==26.1.0
audioread==3.1.0
babel==2.18.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cmudict==1.1.3
colorama==0.4.6
contourpy==1.3.3
csvw==3.7.0
cycler==0.12.1
decorator==5.2.1
diskcache==5.6.3
distro==1.9.0
dlinfo==2.0.0
espeakng==1.0.3
fastapi==0.135.1
fonttools==4.62.1
greenlet==3.3.2
groq==1.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==9.0.0
importlib_resources==6.5.2
iniconfig==2.3.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
kneed==0.8.6
language-tags==1.2.0
lazy-loader==0.5
librosa==0.11.0
llama_cpp_python==0.3.16
llvmlite==0.46.0
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
msgpack==1.1.2
numba==0.64.0
numpy==2.4.3
openai==2.29.0
packaging==26.0
pandas==3.0.1
phonemizer==3.3.0
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
pooch==1.9.0
praatio==6.2.2
pronouncing==0.2.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
rdflib==7.6.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rfc3986==1.5.0
rich==14.3.3
rich-click==1.9.7
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
segments==2.4.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
SQLAlchemy==2.0.48
starlette==1.0.0
termcolor==3.3.0
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.42.0
win32_setctime==1.2.0
zipp==3.23.0
